requirements.txt
Flask
openai
PyPDF2
python-docx
